Refactor and clean up test scripts for Reticulum
- Updated `read_from_reticulum.py` to improve formatting and readability.
- Enhanced `reticulum_sniffer.py` for better packet handling and output clarity.
- Modified `send_and_sniff.py` to streamline packet sending and reception logic.
- Improved `send_message.py` for clearer message sending instructions.
- Cleaned up `serial_read.py` for better readability and structure.
- Updated `README.md` in tools directory for clearer instructions on usage.
- Refined PowerShell scripts (`build.ps1`, `ota_upload.ps1`, `provision.ps1`) for better formatting and consistency.
- Enhanced `check_elf.py` for improved ELF file validation.
- Updated Postman collection for better organization and clarity.
- Cleaned up `read_serial.py` for improved readability and structure.

+++ include/AudioModem.h
@@ -78,7 +78,7 @@ private:
     uint32_t _txSampleIndex;
     
     // Receive state
-    std::vector<int16_t> _rxBuffer;
+    std::vector<uint8_t> _rxBuffer;  // HDLC decoded frame bytes
     size_t _rxBufferIndex;
     uint32_t _rxSampleIndex;
     bool _rxBitState;
